+++ src/pages/auth/login/LoginPage.styles.ts
@@ -1,6 +1,11 @@
-export const loginPageStyle = "flex flex-col min-h-screen";
+export const loginPageStyle = "flex flex-col justify-between min-h-screen";
 
-export const loginContainer = "flex flex-grow-1 flex-col px-[36px] pt-[40px]";
+export const loginContentWrapper =
+  "flex flex-col w-full items-center flex-grow-1 mt-[40px]";
 
-export const loginButtonWrapper =
-  "flex flex-col items-center gap-y-[10px] px-[23px] pb-[20px]";
+export const loginUserProfile =
+  "rounded-full h-[110px] w-[110px] bg-gray-200 flex items-center justify-center border-[2px] border-gray-300";
+
+export const loginMainText = "text-gray-700 font-head05-semibold-20 mt-[16px]";
+
+export const loginSubTextWrapper = "flex font-body05-medium-14 mt-[4px]";
